Fixed backwards diagonal algorithm
- Removed comments
- Cleaned modules
- Removed test
