+++ engine.py
@@ -1,3 +1,5 @@
+"""Game Engine Module"""
+
 import terminaltables as tb
 from copy import deepcopy
 
@@ -84,13 +86,13 @@ class TicTacToe:
 
         for x in range(self.size):
             y = (x - r) + c
-            if 0 <= y < self.size:
+            if y in range(self.size):
                 if coordinates:
                     forwards.append((x, y))
                 else:
                     forwards.append(self.get(x, y))
 
-            y = (r + x) + c
+            y = (r - x) + c
             if 0 <= y < self.size:
                 if coordinates:
                     backwards.append((x, y))
@@ -190,181 +192,6 @@ class TicTacToe:
             del self.snapshots[index]
         except IndexError:
             pass
-
-class UltimateTicTac(TicTacToe):
-    """Ultimate TicTacToe Board Class"""
-
-    def __init__(self, size=3):
-        """Initialize board matrix
-        
-        Arguments:
-            size {int} -- length of board sides (default: {3})
-        """
-        self.matrix = [[[[' ']*size for _c in range(size)] for _r in range(size)] for _r in range(size)]
-
-        # Master board, summary of inner board states
-        self.gist = [[' '] * size for row in range(size)] 
-
-    @property
-    def board(self):
-        """Return a printable representation of the current board state"""
-        _master = []
-        for _m_row in self.matrix:
-            _m_cols = []
-            for _m_col in _m_row:
-                _i_tb = tb.DoubleTable(_m_col)
-                _i_tb.inner_heading_row_border = False
-                _i_tb.inner_row_border = True
-                _m_cols.append(_i_tb.table)
-            _master.append(_m_cols)
-        _m_tb = tb.DoubleTable(_master)
-        _m_tb.inner_heading_row_border = False
-        _m_tb.inner_row_border = True
-        return _m_tb.table
-
-    @board.deleter
-    def board(self):
-        """Reset board"""
-        self.matrix = [[[[' ']*size for _c in range(size)] for _r in range(size)] for _r in range(size)]
-
-        # Master board, summary of inner board states
-        self.gist = [[' '*size for _r in range(size)]]
-
-    def row(self, n, board=None):
-        """Get row values
-        
-        Arguments:
-            n {int} -- zero-based row number
-            board {tuple} -- (r, c) coordinates of inner board (default: {(0,0)}); set None to get from gist board
-        
-        Returns:
-            tuple -- elements of the nth row
-        """
-        if board == None:
-            return tuple(self.gist[n])
-        else:
-            br, bc = board
-            return tuple(self.matrix[br][bc][n])
-
-    def column(self, n, board=None):
-        """Get column values
-        
-        Arguments:
-            n {int} -- zero-based column number
-            board {tuple} -- (r, c) coordinates of inner board (default: {(0,0)}); set None to get from gist board
-        
-        Returns:
-            tuple -- elements of the nth column
-        """
-        if board == None:
-            return tuple([row[n] for row in self.gist])
-        else:
-            br, bc = board
-            return tuple([row[n] for row in self.matrix[br][bc]])
-
-    def diagonals(self, r, c, board=None, coordinates=False):
-        """Get diagonal values intersecting (r, c) coodinates
-
-        Parameters:
-            r {int} -- a zero based row number
-            c {int} -- a zero based column number
-            board {tuple} -- (r, c) coordinates of inner board (default: {(0,0)}); set None to get from gist board
-            coodinates {bool} -- returns cell coordinates instead of values
-
-        Returns:
-            dict -- forward and backward diagonal values
-        """
-        forwards = []
-        backwards = []
-
-        for x in range(self.size):
-            y = (x - r) + c
-            if 0 <= y < self.size:
-                if coordinates:
-                    forwards.append((x, y))
-                else:
-                    forwards.append(self.get(x, y, board))
-
-            y = (r + x) + c
-            if 0 <= y < self.size:
-                if coordinates:
-                    backwards.append((x, y))
-                else:
-                    backwards.append(self.get(x, y, board))
-
-        return {'forward': tuple(forwards), 'backward': tuple(backwards)}        
-
-    def get(self, r, c, board=None):
-        """Get cell value
-
-        Arguments:
-            r {int} -- zero-based cell row number
-            c {int} -- zero-based cell column number
-            board {tuple} -- (r, c) coordinates of inner board (default: {(0,0)}); set None to get from gist board
-            force {bool} -- overwrite value on non-empty cells
-
-        Returns:
-            mixed -- value {str}, False {bool} if (r, c) does not exist
-        """
-        if 0 <= r < self.size and 0 <= c < self.size:
-            if board == None:
-                return self.gist[r][c]
-            else:
-                br, bc = board
-                return self.matrix[br][bc][r][c]
-        else:
-            return False
-
-    def set(self, value, r, c, board=None, force=False):
-        """Sets the cell value based on (x, y, ix, iy) coordinates or just on (x, y) coordinates
-        
-        Arguments:
-            value {str}
-            r {int} -- zero-based cell row number
-            c {int} -- zero-based cell column number
-            board {tuple} -- (r, c) coordinates of inner board (default: {(0,0)}); set None to get from gist board
-
-        Returns:
-            bool -- True on success, False on fail
-        """
-        if 0 <= r < self.size and 0 <= r < self.size:
-            if board == None:
-                if self.get(r, c) == ' ' or force:
-                    self.gist[r][c] = value
-                    return True
-            else:
-                br, bc = board
-                if self.get(r, c, (br, bc)) == ' ' or force:
-                    self.matrix[br][bc][r][c] = value
-                    return True
-        else:
-            return False
-
-    def check(self, r, c, board=None):
-        """Check if (r, c) value exists at least three times consecutively lines in any direction (horizontal, vertical, diagonal)
-        intersecting with (r, c)
-        
-        Arguments:
-            r {int} -- a zero based row number
-            c {int} -- a zero based column number
-            board {tuple} -- (r, c) coordinates of inner board (default: {(0,0)}); set None to get from gist board
-        
-        Returns:
-            bool
-        """
-        value = self.get(r, c, board)
-
-        if str(value * 3) in ''.join(self.row(r, board)):
-            return True
-
-        if str(value * 3) in ''.join(self.column(c, board)):
-            return True            
-
-        for direction, values in self.diagonals(r, c, board).items():
-            if str(value * 3) in ''.join(values):
-                return True
-                
-        return False
 
 class FlipTacToe(TicTacToe):
     """FlipTacToe Board Class"""
